Share temp copy and diagnostic remap helpers
